code/requirements.txt
Flask==2.3.2
wechatpy==1.8.18
cryptography
redis==4.5.5
python-redis-lock==4.0.0
gunicorn==20.1.0
gevent